Remove rules from Slugify and move them into files
There is a generate-rules.php script that reads the
JSON files and updates a DefaultRuleProvider which allows
Slugify to still be very fast.

In addition this PR cleans up a lot of stuff, this is required to
keep the code clean, maintainable, readable and testable when we
add features in the future.